Unified the test_3m and test_threem test files in test_bibliotheca.
